Un pequenio avanze
Se anaden, el xml de categorias, la clase categorias, y se actualizan el lector y escritor para soportar categorias

Seguido se editaran el lector y escritor para soportar Soportes

+++ src/aplicacion/Cliente.java
@@ -39,7 +39,7 @@ public class Cliente {
 		
 	}
 	
-	public void agregarPaquete(String trackingID, String tienda, String courier, int valor, String descripcion) {
+	public void agregarPaquete(String trackingID, String tienda, String courier, double valor, String descripcion) {
 		Paquete paqueteNuevo = new Paquete(trackingID,tienda,courier,valor,descripcion);
 		paquetes.add(paqueteNuevo);
 	}
